[Fix] Faq User, Admin 기능 분리
